Some changes to support more backends.
Refactor the testing backend.
Update and clean up the framework.
Many fixes to support the xdp target.
