+++ backends/ebpf/ebpfBackend.cpp
@@ -37,10 +37,10 @@ void run_ebpf_backend(const EbpfOptions& options, const IR::ToplevelBlock* tople
     }
 
     Target* target;
-    if (options.target.isNullOrEmpty() || options.target == "bcc") {
-        target = new BccTarget();
-    } else if (options.target == "kernel") {
+    if (options.target.isNullOrEmpty() || options.target == "kernel") {
         target = new KernelSamplesTarget();
+    } else if (options.target == "bcc") {
+        target = new BccTarget();
     } else if (options.target == "test") {
         target = new TestTarget();
     } else {
