+++ testdata/p4_16_samples/hit_ebpf.stf
@@ -17,4 +17,3 @@ packet 1 001b1700 0130b881 98b7aeb7 08004500 00344a6f 40004006 53920a01 98463212
 
 packet 1 001b1700 0130b881 98b7aeb7 08004500 00344a6f 40004006 53920a01 98453212 c86acf2c 01bbd0fa 585c4ccc b2ac8010 0353c314 00000101 080a0192 463911a0 c06f
 expect 1 001b1700 0130b881 98b7aeb7 08004500 00344a6f 40004006 53920a01 98453212 c86acf2c 01bbd0fa 585c4ccc b2ac8010 0353c314 00000101 080a0192 463911a0 c06f
-
